remove enclosure
