--- mycroft/client/enclosure/version.txt
@@ -1,1 +1,0 @@
-1.4.0
