+++ mycroft/configuration/mycroft.conf
@@ -378,27 +378,6 @@
         }
   },
 
-  // Mark 1 enclosure settings
-  // Override: SYSTEM (e.g. Picroft)
-  "enclosure": {
-    // Platform name
-    // Options: 'picroft', 'mycroft_mark_1'
-    // Override: SYSTEM (set by specific enclosures)
-    # "platform": "picroft",
-    # "platform_enclosure_path": "/etc/myenclosure/code.py",
-
-    // COMM params to the Arduino/faceplate
-    "port": "/dev/ttyAMA0",
-    "rate": 9600,
-    "timeout": 5.0,
-
-    // ??
-    "update": true,
-
-    // Run a self test at bootup?
-    "test": false
-  },
-
   // Level of logs to store, one of  "CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG"
   // NOTE: This configuration setting is special and can only be changed in the
   // SYSTEM or USER configuration file, it will not be read if defined in the
