+++ mycroft/enclosure/__main__.py
@@ -1,0 +1,10 @@
+from mycroft.enclosure.gui import GUIManager
+
+
+def main():
+    gui = GUIManager()
+    gui.run()
+
+
+if __name__ == "__main__":
+    main()
