+++ setup.py
@@ -68,7 +68,7 @@ setup(
             'mycroft-audio=mycroft.audio.__main__:main',
             'mycroft-echo-observer=mycroft.messagebus.client.ws:echo',
             'mycroft-audio-test=mycroft.util.audio_test:main',
-            'mycroft-enclosure-client=mycroft.client.enclosure.__main__:main',
+            'mycroft-gui-listener=mycroft.enclosure.__main__:main',
             'mycroft-cli-client=mycroft.client.text.__main__:main'
         ]
     }
